+++ internal/format/presenter.go
@@ -3,7 +3,6 @@ package format
 import (
 	"github.com/wagoodman/go-presenter"
 
-	"github.com/anchore/grype/grype/presenter/csv"
 	"github.com/anchore/grype/grype/presenter/cyclonedx"
 	"github.com/anchore/grype/grype/presenter/json"
 	"github.com/anchore/grype/grype/presenter/models"
@@ -27,7 +26,7 @@ func GetPresenter(format Format, c PresentationConfig, pb models.PresenterConfig
 	case TableFormat:
 		return table.NewPresenter(pb, c.ShowSuppressed)
 	case CSVFormat:
-		return csv.NewPresenter(pb, c.ShowSuppressed)
+		return table.NewPresenterWithOptions(pb, c.ShowSuppressed, true)
 
 	// NOTE: cyclonedx is identical to EmbeddedVEXJSON
 	// The cyclonedx library only provides two BOM formats: JSON and XML
